Refactor: Extract Canvas Interactions to Controller
