+++ RailmlEditor.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RailmlEditor", "RailmlEditor\RailmlEditor.csproj", "{0299C39B-5619-4D23-A82D-3C378BC30157}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RailmlEditor.Tests", "RailmlEditor.Tests\RailmlEditor.Tests.csproj", "{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,18 @@ Global
 		{0299C39B-5619-4D23-A82D-3C378BC30157}.Release|x64.Build.0 = Release|Any CPU
 		{0299C39B-5619-4D23-A82D-3C378BC30157}.Release|x86.ActiveCfg = Release|Any CPU
 		{0299C39B-5619-4D23-A82D-3C378BC30157}.Release|x86.Build.0 = Release|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Debug|x64.Build.0 = Debug|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Debug|x86.Build.0 = Debug|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Release|x64.ActiveCfg = Release|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Release|x64.Build.0 = Release|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Release|x86.ActiveCfg = Release|Any CPU
+		{FFC77EC6-AF1F-4746-8EE9-EC85601E12B7}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
